ollama-langchain-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.7
ollama==0.1.7
requests==2.31.0
python-multipart==0.0.6
